Created A Skeleton Web Server to build off of.
This is a skeleton web server with basic structure to get us started.

The frontend files/applications should remain in the views directory.
Sub-directories can and should be made similar to the test2-views
directory. This can keep our frontend code organized and logically
seperated.

Any public assets like css stylesheets, images, and javascript files
that need to be accessed should be in the public folder. You can access
these files with the server url and file name i.e. http://localhost:3000/orang.jpg
for the orangutan image. Sub folders can and should be made in this
folder as necessary.

Any and all additional files for routing should be created and placed in
the routes folder, similar to test2.js. Files in the routes folder
should only provide routing. If significant amounts of code is necessary
to be paired with a specific route, it may be better to create a
seperate directory for middleware and put the function in a file in that
directory.