Login + Registration from the Backend (#8)
* Login + Registeration from the Backend

This commit further seperates the backend server from the frontend, implementing a model - routes - controllers structure where the models are database models/schema, routes are the backend api endpoints, and the controllers are the business logic such as the registration feature.

The main features provided in this commit are user registration, user login, JWT token regeneration with refresh tokens, and the ability to confirm your email address through an email sent to you(you have to construct a post request with curl or postman as the link isn't set up, the backend works though).

Additionally, the envExample for the backend was modified to include new required variables. These include the SECRET_ACCESS_TOKEN/EXPIRY which are the secret key needed to generate JWT access token and the time they expire. Also there is the SECRET_REFRESH_TOKEN/EXPIRY which are the secret key needed to generate refresh tokens, and the time that those tokens expire. Additionaly there are seperate DB environment vars that are used fro constructed the DB access link. Lastly there is the EMAIL_USER/PASSWORD vars which are the credentials needed to access the emailer.

The functionalities have been tested in post man prior to commiting here.

* Whoopsie, these sllipped through

* Tragic Semicolon Issues Resolved

I refuse to accept blame on principle, because I am perfect.